# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jobseeker",
    "description": "Self-hosted job-search workbench",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/pair": {
      "post": {
        "tags": [],
        "operationId": "auth_pair",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [],
        "operationId": "events",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "Replay events with id greater than this",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "text/event-stream of DomainEvent"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PairRequest": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
...
    }
  }
}